+++ src/main/java/org/zhadok/poe/controller/action/ActionHandlerMacro.java
@@ -3,7 +3,7 @@ package org.zhadok.poe.controller.action;
 import java.awt.Robot;
 
 import org.zhadok.poe.controller.action.macro.Macro;
-import org.zhadok.poe.controller.config.pojo.ConfigAction;
+import org.zhadok.poe.controller.config.pojo.Mapping;
 
 import net.java.games.input.Event;
 
@@ -14,17 +14,18 @@ public class ActionHandlerMacro extends ActionHandler {
 	
 	public ActionHandlerMacro(Robot robot, ActionHandlerKey actionHandlerKey, ActionHandlerMouse actionHandlerMouse) {
 		super(robot);
+		log(1, "Initializing ActionHandlerMacro..."); 
 		this.actionHandlerKey = actionHandlerKey; 
 		this.actionHandlerMouse = actionHandlerMouse; 
 	}
 
 	@Override
-	public void handleAction(Event event, ConfigAction action) {
-		
+	public void handleAction(Event event, Mapping mapping) {
+		log(3, "Handling event with mapping=" + mapping.toString()); 
 		if (event.getValue() != 0) {
 			try {
-				Macro macro = Macro.buildMacro(action.getMacro(), robot);
-				macro.startMacro(event, action.getMacro(), this);
+				Macro macro = Macro.buildMacro(mapping.getAction().getMacro(), robot);
+				macro.startMacro(event, mapping.getAction().getMacro(), this);
 			}
 			catch(Exception e) {
 				e.printStackTrace();