+++ src/main/java/org/zhadok/poe/controller/config/ui/NextControllerMappingHandler.java
@@ -1,5 +1,0 @@
-package org.zhadok.poe.controller.config.ui;
-
-public class NextControllerMappingHandler {
-
-}